Start queue_main refactoring for testability (#9501)
- Move proxy handler to its own file.
- Move benchmark and one test that seemed to be only proxy related.
- Add a new test to cover kubelet bypass

More to come!
