+++ cmd/queue/main.go
@@ -48,7 +48,6 @@ import (
 	"knative.dev/pkg/tracing"
 	tracingconfig "knative.dev/pkg/tracing/config"
 	"knative.dev/pkg/tracing/propagation/tracecontextb3"
-	"knative.dev/serving/pkg/activator"
 	activatorutil "knative.dev/serving/pkg/activator/util"
 	pkghttp "knative.dev/serving/pkg/http"
 	"knative.dev/serving/pkg/http/handler"
@@ -101,55 +100,6 @@ type config struct {
 	TracingConfigSampleRate           float64                   `split_words:"true"` // optional
 	TracingConfigZipkinEndpoint       string                    `split_words:"true"` // optional
 	TracingConfigStackdriverProjectID string                    `split_words:"true"` // optional
-}
-
-// Make handler a closure for testing.
-func proxyHandler(breaker *queue.Breaker, stats *network.RequestStats, tracingEnabled bool, next http.Handler) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if network.IsKubeletProbe(r) {
-			next.ServeHTTP(w, r)
-			return
-		}
-
-		if tracingEnabled {
-			proxyCtx, proxySpan := trace.StartSpan(r.Context(), "queue_proxy")
-			r = r.WithContext(proxyCtx)
-			defer proxySpan.End()
-		}
-
-		// Metrics for autoscaling.
-		in, out := network.ReqIn, network.ReqOut
-		if activator.Name == network.KnativeProxyHeader(r) {
-			in, out = network.ProxiedIn, network.ProxiedOut
-		}
-		stats.HandleEvent(network.ReqEvent{Time: time.Now(), Type: in})
-		defer func() {
-			stats.HandleEvent(network.ReqEvent{Time: time.Now(), Type: out})
-		}()
-		network.RewriteHostOut(r)
-
-		// Enforce queuing and concurrency limits.
-		if breaker != nil {
-			var waitSpan *trace.Span
-			if tracingEnabled {
-				_, waitSpan = trace.StartSpan(r.Context(), "queue_wait")
-			}
-			if err := breaker.Maybe(r.Context(), func() {
-				waitSpan.End()
-				next.ServeHTTP(w, r)
-			}); err != nil {
-				waitSpan.End()
-				switch err {
-				case context.DeadlineExceeded, queue.ErrRequestQueueFull:
-					http.Error(w, err.Error(), http.StatusServiceUnavailable)
-				default:
-					w.WriteHeader(http.StatusInternalServerError)
-				}
-			}
-		} else {
-			next.ServeHTTP(w, r)
-		}
-	}
 }
 
 func knativeProbeHandler(logger *zap.SugaredLogger, healthState *health.State, prober func() bool, isAggressive bool, tracingEnabled bool, next http.Handler) http.HandlerFunc {
@@ -380,7 +330,7 @@ func buildServer(ctx context.Context, env config, healthState *health.State, rp 
 	if metricsSupported {
 		composedHandler = requestAppMetricsHandler(logger, composedHandler, breaker, env)
 	}
-	composedHandler = proxyHandler(breaker, stats, tracingEnabled, composedHandler)
+	composedHandler = queue.ProxyHandler(breaker, stats, tracingEnabled, composedHandler)
 	composedHandler = queue.ForwardedShimHandler(composedHandler)
 	composedHandler = handler.NewTimeToFirstByteTimeoutHandler(composedHandler, "request timeout", handler.StaticTimeoutFunc(timeout))
